Created a VndbUtils class for retrieving the dumps. I created a 60 second limit in between requests, so programmers/users can't request database dumps in quick sucession. I may need to change the response message on it though.
Set up an option of using a custom userAgent when sending the request.
Also removed the bytesRead that wasn't being used.
